status.utils - Basic support for pynag MultiSite
This patch changes status.utils.livestatus() to use
the new pynag.Parsers.MultiSite() class.

There is still a lot of things to do. But this patch is a good start.

Parsers.MultiSite maintains backwards compatibility with old mk_livestatus
class so that most views in Adagios should still work.

What is left:
* Go through each view, and make sure that all calls to specific
  nagios object have the backend="something" so when a user clicks a
  service is definately loaded from the correct backend (in case of naming conflicts)

* Make a settings dialog for this (currently it is assumed that livestatus_path in adagios.conf
  is a seperated list, this is not ideal and might not be the case in the future.