Exception handling around volatite userdata
This patch makes adagios a little more elegant
when there are problems getting the user-specific data.

Specifically, it will display a warning instead of a traceback
if userprefs is the only problem we are facing.

Additionally fix a problem where stylesheets were not correctly
loaded if user prefs had problem.

Fixes #366