service_detail page javascript errors on spaces
This patch fixes a bug in adagios.misc.populate_search_with_querystring_fields() javascript code
that would generate an incorrect jquery selector if you were on a service detail for a serice
with space in the description.

This would cause javascripts to fail on that page, including 'other services on this host' breaking.