Mail objectlists should link to problems page
Before this patch every time a user would send mail regarding
specific objects, there would be a link at the top of the email to
current page the user was on.

Does not really make since, for example if you acknowledge hosts from
within the unhandled problems page, to be linked to the unhandled problems
page.

Therefore, changing this link to always point to the problems page.

Fixes #501