Update rest.py
In Naemon 0.9.1 the service recheck option fails without error.  It says the command is sent successfully, but naemon.log shows the following:
Parse error: Invalid characters (.78) in ulong '1426878227.78'

The Nagios example (http://old.nagios.org/developerinfo/externalcommands/commandinfo.php?command_id=129) shows the use of `date +%s` to populate the time, which is an integer, so decimal is not likely expected.  Although Nagios 3.5.1 seems to be okay with this, Naemon does strong type checking and rejects it.