Fix directory for Git Event Handler. (EventHandlerError: fatal: Not a git repository (or any of the parent directories): .git)
When you enable Git support in Adagios, AND are using nonstandard
directory base (e.g. /etc/shinken instead of /etc/nagios), it
is not possible to create / edit configuration.
More specifically, each time Git Handler wants to commit something,
the Web UI fails with the following error:
EventHandlerError: fatal: Not a git repository (or any of the parent directories): .git

Root cause seems to be that the directory with which Git Handler
is configured is not the proper one, but a default one ( /etc/nagios ),
which is not under git control.

This (simple) fix aims to correct this.